refactor

+++ src/standard.rs
@@ -95,7 +95,7 @@ impl StandardSolitaire {
 
     #[must_use]
     pub fn find_deck_card(&self, card: &Card) -> Option<u8> {
-        for i in 0..N_FULL_DECK as u8 {
+        for i in 0..N_FULL_DECK {
             let offset = self.deck.offset(i);
             if offset > 0 && self.deck.peek(offset - 1) == card {
                 return Some(i);
